added appending `__version` to worklets removed plugin version check moved cpp version check back to original file modified tests and plugin to be able to run tests
